requirements.txt
streamlit
google-generativeai
pdfplumber
matplotlib
pandas
seaborn
numpy
pillow